requirements.txt
tensorflow==2.4.1
matplotlib==3.5.1
torch==1.8.1
pandas==1.2.4
seaborn==0.11.1
opencv_python_headless==4.5.5.62
Flask==1.1.2
onnxruntime==1.10.0
easyocr==1.4.1
coremltools==5.2.0
torchvision==0.9.1
requests==2.25.1
psutil==5.8.0
numpy==1.16.4
scipy==1.4.1
tqdm==4.63.0
openvino==2021.4.2
pytesseract==0.3.9
albumentations==1.1.0
export==0.2.0
ipython==8.3.0
keras_ocr==0.8.9
pafy==0.5.5
Pillow==9.1.0
PyYAML==6.0
tensorrt==0.0.1.dev5

